Split semian operations into more cohesive translation units
