+++ .travis.yml
@@ -3,6 +3,7 @@ language: ruby
 sudo: true
 
 before_install:
+  - gem update --system
   - gem install bundler
   - scripts/install_toxiproxy.sh
