+++ ext/semian/semian.h
@@ -1,0 +1,34 @@
+/*
+System, 3rd party, and project includes
+
+Implements Init_semian, which is used as C/Ruby entrypoint.
+*/
+
+#ifndef SEMIAN_H
+#define SEMIAN_H
+
+// System includes
+#include <sys/types.h>
+#include <sys/ipc.h>
+#include <sys/sem.h>
+#include <sys/time.h>
+#include <errno.h>
+#include <string.h>
+#include <stdio.h>
+
+// 3rd party includes
+#include <openssl/sha.h>
+#include <ruby.h>
+#include <ruby/util.h>
+#include <ruby/io.h>
+
+//semian includes
+#include <semian_types.h>
+#include <semian_resource.h>
+#include <semian_resource_alloc.h>
+#include <semian_tickets.h>
+#include <semset.h>
+
+void Init_semian();
+
+#endif //SEMIAN_H
