+++ ext/semian/semian_resource_alloc.c
@@ -1,0 +1,12 @@
+#include <semian_resource_alloc.h>
+
+const rb_data_type_t
+semian_resource_type = {
+  "semian_resource",
+  {
+    semian_resource_mark,
+    semian_resource_free,
+    semian_resource_memsize
+  },
+  NULL, NULL, RUBY_TYPED_FREE_IMMEDIATELY
+};
